requirements.txt
apache-airflow-providers-google==6.1.0
requests==2.25.1
google-cloud-bigquery==3.0.0
dbt-bigquery==1.0.0